refactor: PathNavigationTest 테스트 코드 추가요금 테스트 코드 분리
